Basic Enemy AI (Without Damage integrated yet) (#49)
* Beginning Stinger AI Development

Created a EnemyDevelopmentScene to use with the development of the Stinger (and any other) enemy, the StingerSuperClass.cs script, and the EnemyHealth.cs script.

* Update EnemyDevelopmentScene

Just saved it, nothing much.

* Working on the AI Movement (Stinger Enemy)

Having the Enemy chase the player, trying to restrict the Stinger/enemy's movement only to the y-axis (for now).

* New Changes to Enemy Stinger Superclass

WIP

* Created Prefabs Folder and Continued work on Enemy AI

Got the movement working, for now. Worried about the enemies bunching up later.  Using raycasting to detect if the player is in site, now just working on actually making a bullet fire.

* Committing Updated Progress on Enemy AI

Created StingerBullet prefab and StingerBullet.cs, got the shooting mechanics of the Stinger gameobject working. Rearranged components on the Player game object in the PlayerDevelopmentScene. Based the Player gameObject weight in the PlayerDevelopmentScene off of the F-18A Hornet and the StingerBullet is set to 0.5 kilograms.

* Commented Code and Deleted EnemyHealth.cs

Commented the code in StingersuperClass.cs and deleted EnemyHealth.cs from the project.