Made Level & Implimented Models (#50)
I made a scene as a level where everyone can play the game. We will use this scene for initialization so be careful about editing it. I also implimented some FBX models into Unity. We will be using the FBX models for all models in the game. We will NOT be using the .mb files, only the .fbx files as some machines won't render the .mb files properly and Unity supports .fbx files natively.